# backend/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: GreeterService API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /api/v1/hello/{name}:
        get:
            tags:
                - GreeterService
            description: Sends a greeting
            operationId: GreeterService_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /api/v1/query:
        get:
            tags:
                - GreeterService
            operationId: GreeterService_Query
            parameters:
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.QueryReply'
components:
    schemas:
        helloworld.v1.Data:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                username:
                    type: string
                age:
                    type: string
                gender:
                    type: integer
                    format: enum
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        helloworld.v1.QueryReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/helloworld.v1.Data'
                code:
                    type: integer
                    format: uint32
                message:
                    type: string
tags:
    - name: GreeterService